backend/requirement.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
alembic==1.16.2
pytest==8.4.1
httpx==0.28.1
python-multipart==0.0.20
